Refactor(minjae): 코드 리팩토링 및 Scss 파일 및 폴더 구조화 #28
