--- src/App.scss
@@ -1,75 +1,0 @@
-//Header
-@import './scss/Header/Header';
-
-//HomePage
-@import './scss/HomePage/MainBanner';
-@import './scss/HomePage/Funtion';
-@import './scss/HomePage/Introduce';
-@import './scss/HomePage/Notice';
-
-//Login, Register
-@import './scss/LoginPage/Login';
-@import './scss/RegisterPage/Register';
-
-* {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-  font-family: 'Poppins', sans-serif;
-  scroll-behavior: smooth;
-}
-
-/* a 링크 초기화 */
-a {color: #222; text-decoration: none;}
-a:hover {color: #2698cb;}
-
-/* 블릿기호 초기화 */
-ul, li, ol {list-style: none;}
-
-/* 제목 태그 초기화 */
-h1, h2, h3, h4, h5, h6 {font-size: 13px; font-weight: normal;}
-
-/* 버튼 초기화*/
-button{ border: 0;}
-
-/* clearfix */
-.clearfix:before, .clearfix::after{display: block; content: ""; line-height: 0;}
-.clearfix::after {clear: both;}
-.clearfix{ *zoom: 1;}
-
-/* input style */
-input {background-image: none;}
-input[type=button],
-input[type=reset],
-input[type=text],
-input[type=password],
-input[type=submit],
-input[type=search],
-input[type=tel],
-input[type=email] {
-    -webkit-appearance: none;
-    border-radius: 0
-}
-input[type=search]::-webkit-search-cancel-button,
-input[type=search]::-webkit-search-decoration {
-    -webkit-appearance: none
-}
-input:checked[type=checkbox] {
-    background-color: #666;
-    -webkit-appearance: checkbox
-}
-
-/* IR 효과 */
-.ir_pm {
-    display:block; overflow:hidden; font-size:0; line-height:0; text-indent:-9999px;
-} 
-/* 의미있는 이미지의 대체 텍스트를 제공하는 경우(Phark Method) */
-.ir_wa {
-    display:block; overflow:hidden; position:relative; z-index:-1; width:100%; height: 100%;
-} 
-/* 의미있는 이미지의 대체 텍스트로 이미지가 없어도 대체 텍스트를 보여주고자 할 때(WA IR) */
-.ir_su {
-    overflow: hidden; position:absolute; width:0; height:0; line-height:0; text-indent:-9999px;
-}
- /* 대체 텍스트가 아닌 접근성을 위한 숨김 텍스트를 제공할 때 */
-
