+++ src/pages/RegisterPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Header from '../components/Header/Header';
-import RegisterComponent from '../components/Register/RegisterComponent';
+import RegisterComponent from '../components/auth/RegisterComponent';
 
 const RegisterPage = () => {
     return (
