--- src/scss/LoginPage/Login.scss
@@ -1,142 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Montserrat&display=swap');
-
-#login {
-    width: 100%;
-    height: 100vh;
-    background-color: #eee5de;
-
-    .container {
-        width: 100%;
-        height: 100vh;
-        display: flex;
-        justify-content: center;
-        flex-direction: column;
-        align-items: center;
-    }
-}
-
-.login-form{
-    width: 400px;
-    background-color: #fdfdfd;
-    border-radius: 10px;
-    padding: 0 40px;
-
-    .login-title {
-        text-align: center;
-        padding: 40px 0 10px 0;
-        font-size: 2rem;
-        font-family: 'Fredoka One', cursive;
-        color: #222;
-    }
-    
-    .login-text-field {
-        position: relative;
-        border-bottom: 2px solid #adadad;
-        margin: 40px 0;
-
-        input {
-            font-family: 'Montserrat', sans-serif;
-            width: 100%;
-            padding: 0 5px;
-            height: 40px;
-            outline: none;
-            border: none;
-            font-size: 16px;
-
-            &:focus ~ label {
-                top: -5px;
-                color: #d3a887;
-            }
-
-            &:focus ~ span::before {
-                width: 100%;
-            }
-        }
-        
-        label {
-            font-family: 'Montserrat', sans-serif;
-            position: absolute;
-            top: 50%;
-            left: 5px;
-            color: #adadad;
-            transform: translateY(-50%);
-            font-size: 16px;
-            pointer-events: none;
-            transition: .5s top;
-            &.fix {
-                top: -5px;
-                color: #d3a887;
-            }
-        }
-
-        span {
-            &::before {
-                content: '';
-                position: absolute;
-                top: 40px;
-                left: 0;
-                width: 0;
-                height: 2px;
-                background-color: #cc8e5e;
-                transition: 1s;
-            }
-            &.input-border::before {
-                width: 100%;
-            }
-        }
-    }
-    
-    .search {
-        font-size: 0.8rem;
-        a {
-            margin: 0 2px;
-            &:hover {
-                color: #d3a887;
-                text-decoration: underline;
-            }
-        }
-        p { display: inline-block; }
-    }
-
-    input[type="submit"] {
-        font-family: 'Montserrat', sans-serif;
-        width: 100%;
-        height: 50px;
-        border: 1px solid;
-        background: #d3a887;
-        border-radius: 25px;
-        font-size: 1.2rem;
-        color : #fdfdfd;
-        font-weight: 700;
-        cursor: pointer;
-        outline: none;
-        margin: 20px 0 10px 0;
-    }
-
-    input[type="submit"]:hover {
-        background: rgb(204, 154, 115);
-        transition: .5s;
-    }
-
-    .signup-link {
-        font-family: 'Montserrat', sans-serif;
-        margin: 20px 0;
-        text-align: center;
-        font-size: 16px;
-        color: #666666;
-        
-        a {
-            font-family: 'Montserrat', sans-serif;
-            color: #d3a887;
-            text-decoration: none;
-
-            &:hover { text-decoration: underline; }
-        }
-    }
-    .login-err {
-        color:red;
-        font-size: 0.8rem;
-        text-align: center;
-        margin-bottom: 15px;
-    }
-}
