+++ src/scss/layout/_forms.scss
@@ -1,0 +1,155 @@
+.form-template {
+    width: 100%;
+    height: 100vh;
+    background-color: #eee5de;
+
+    .container {
+        width: 100%;
+        height: 100vh;
+        display: flex;
+        justify-content: center;
+        flex-direction: column;
+        align-items: center;
+    }
+}
+
+.form-field {
+    width: 400px;
+    background-color: #fdfdfd;
+    border-radius: 10px;
+    padding: 0 40px;
+
+    .form-title {
+        font-family: 'Fredoka One', cursive;
+        text-align: center;
+        padding: 40px 0 10px 0;
+        font-size: 2rem;
+        color: #222;
+    }
+}
+
+.text-field {
+    position: relative;
+    border-bottom: 2px solid #adadad;
+    margin: 40px 0;
+
+    .input-text {
+        font-family: 'Montserrat', sans-serif;
+        width: 100%;
+        padding: 0 5px;
+        height: 40px;
+        outline: none;
+        border: none;
+        font-size: 16px;
+
+        &:focus ~ .input-label {
+            top: -5px;
+            color: #d3a887;
+        }
+
+        &:focus ~ .input-border::before {
+            width: 100%;
+        }
+    }
+
+    .input-border {
+        &::before {
+            content: '';
+            position: absolute;
+            top: 40px;
+            left: 0;
+            width: 0;
+            height: 2px;
+            background-color: #cc8e5e;
+            transition: 1s;
+        }
+        &.fill::before {
+            width: 100%;
+        }
+    }
+    
+    .input-label {
+        font-family: 'Montserrat', sans-serif;
+        position: absolute;
+        top: 50%;
+        left: 5px;
+        color: #adadad;
+        transform: translateY(-50%);
+        font-size: 16px;
+        pointer-events: none;
+        transition: .5s top;
+        &.fix {
+            top: -5px;
+            color: #d3a887;
+        }
+    }
+}
+
+.form-errmsg {
+    color:red;
+    font-size: 0.8rem;
+    text-align: center;
+    margin-bottom: 15px;
+}
+
+.form-search {
+    font-size: 0.8rem;
+    a {
+        margin: 0 2px;
+        &:hover {
+            color: #d3a887;
+            text-decoration: underline;
+        }
+    }
+}
+
+.form-submit {
+    font-family: 'Montserrat', sans-serif;
+    width: 100%;
+    height: 50px;
+    border: 1px solid;
+    background: #d3a887;
+    border-radius: 25px;
+    font-size: 1.2rem;
+    color : #fdfdfd;
+    font-weight: 700;
+    cursor: pointer;
+    outline: none;
+    margin: 20px 0 10px 0;
+    
+    &:hover {
+        background: rgb(204, 154, 115);
+        transition: .5s;
+    }
+}
+
+.form-signup-link {
+    font-family: 'Montserrat', sans-serif;
+    margin: 20px 0;
+    text-align: center;
+    font-size: 16px;
+    color: #666666;
+    
+    a {
+        font-family: 'Montserrat', sans-serif;
+        color: #d3a887;
+        text-decoration: none;
+
+        &:hover { text-decoration: underline; }
+    }
+}
+
+.form-question {
+    margin: 20px 0;
+    text-align: center;
+    p {
+        font-family: 'Montserrat', sans-serif;
+        display: inline;
+    }
+    a { 
+        font-family: 'Montserrat', sans-serif;
+        margin-left: 10px;
+        color: #d3a887;
+        &:hover { text-decoration: underline; }
+    }
+}
