+++ src/scss/layout/_header.scss
@@ -75,15 +75,6 @@
         display: flex;
         justify-content: center;
         align-items: center;
-
-        &:first-child {
-            // border-top-left-radius: 10px;
-            // border-top-right-radius: 10px;
-        }
-        &:last-child {
-            // border-bottom-left-radius: 10px;
-            // border-bottom-right-radius: 10px;
-        }
     }
 
     .header-menu-drop li:last-child { border-bottom: none; }
