+++ src/scss/pages/home/_funtion.scss
