+++ src/scss/pages/home/_introduce.scss
@@ -1,5 +1,3 @@
-@import url('https://fonts.googleapis.com/css2?family=Thasadith&display=swap');
-
 #home-introduce {
     width: 100%;
     height: 100vh;
