+++ src/scss/pages/home/_mainBanner.scss
