+++ src/scss/pages/home/_notice.scss
@@ -1,5 +1,3 @@
-@import url('https://fonts.googleapis.com/css2?family=Thasadith&display=swap');
-
 #home-notice {
     height: 100vh;
     background-color: #e7dfd9;
